Update Task 3- ATM interface
The task provides user with all the ATM functionalities such as amount withdraw, amount deposit, check balance. It consists of 2 classes Bank_acc and ATM. All the functions of ATM are comprised in ATM class in the form of class methods and the acc_ bal variable  is publicly declared in the Bank_acc class which gets modified as per the actions performed.